add configuration via discord id
