+++ WhosTalking/Configuration.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using Dalamud.Configuration;
 using Dalamud.Plugin;
 
@@ -6,11 +7,13 @@ namespace WhosTalking;
 
 [Serializable]
 public sealed class Configuration: IPluginConfiguration {
+
     // the below exist just to make saving less cumbersome
     [NonSerialized]
     private DalamudPluginInterface? pluginInterface;
 
     public string? AccessToken { get; set; }
+    public List<AssignmentEntry> IndividualAssignments { get; set; } = new();
     public int Version { get; set; } = 0;
 
     public void Initialize(DalamudPluginInterface pluginInterface) {
@@ -20,4 +23,9 @@ public sealed class Configuration: IPluginConfiguration {
     public void Save() {
         this.pluginInterface!.SavePluginConfig(this);
     }
+}
+
+public sealed class AssignmentEntry {
+    public string CharacterName { get; set; } = string.Empty;
+    public string DiscordId { get; set; } = string.Empty;
 }
